requirement.txt
autopep8==1.5.7
cachelib==0.4.1
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.18.1
pycodestyle==2.8.0
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.2
WTForms==2.3.3